Update README.md
I am prepping for a hydra like automation for production art generation based on ai, and keywords, to convey concepts and reflect the deluge of infomration from social news aggregates. A few programs caught my eye, but are, sadly not in action, but were nice enough to show the roadmap to their research. This is one of several. Enjoy.